backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
pydantic-settings>=2.6.0
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.18.0
openai>=1.30.0
python-dotenv>=1.0.0
scipy>=1.11.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
httpx>=0.27.0
llama-cpp-python>=0.3.0
